Implement basic plz sync command (#12)
Implement basic plz sync command

The command operates as follows:

* Fetch the review associated with the current commit
* Traverse up the stack by fetching that review's latest revision's parent
* Traverse each of these reviews in reverse order and do:
  - syncReviewWithParent to update the review
  - If an update happened, fetch the updated remote branch
* Update the current HEAD branch to point to the top of the stack.

Basically it works its way up the stack, updating each review upon the last.
In the event that a parent review was merged, rebased, etc. then the child
review will be rebased accordingly.  At the end of this process, the whole
stack of reviews will be up to date on plz.review, and the origin branches in
the local repo will be up to date with GitHub.

What this command does *not* do is incorporate any local changes. I have some
ideas but it's still a little fuzzy how best to handle it in an intuitive way.

plz-review-url: https://plz.review/review/4914