Consider more revisions when determining parent revision
Previously this code just looked for the latest revision with a matching
head commit SHA, but there is a common case when a parent is merged and
branch auto-deletion is enabled where two revisions with the same base
commit SHA will be created, the first having two commits (the parent's
and the child's), and the second having only the child's. Both of these
will have the same head commit SHA, but only the one where the base
commit SHA matches the parent commit's SHA is relevant.