Handle dropping the bottom commit in a stack
In the event that the bottom commit in a stack is dropped, we should not
follow its linked revisions. Previously we were using it as the starting
point for traversing the linked revisionns because it was the parent of
the latest revision of the bottom review. With this change we specially
handle this situation.