PLZ-1095 Mitigate a race condition creating child revisions
When running `plz review` from the CLI, if plz.review processes the
webhook for a child review's new revision before the webhook for its
parent then the child will be orphaned, and the stack relationship will
be broken. This is a hack to reduce the likelihood that a new revision
in a child review is processed before a new revision in the parent
review. The proper fix is to wait until any revisions that may be
created by pushing the branch or updating the PR are created (e.g. by
polling the API) before continuing on up the stack.

plz-review-url: https://plz.review/review/5606