SSH Handshake Fix (#24)
This fix seemed to have a few components, but ultimately the bug fix merged into go-git solved the reported issue of "ssh: handshake failed: knownhosts: key mismatch". After upgrading this led to a secondary issue of "ssh: handshake failed: ssh: unable to authenticate, attempted methods [none publickey], no supported methods remain" which was solved by upgrading the indirect dependencies crypto and sys as well as running a command which is now recorded in the README.

Unsure if noting known setup issues in the main README is the correct course here.